last_object_clicked Marker Uncaught Type Error
Hi, I was trying to make a project using the bidirectional component such that markers when clicked would send data back for my use in streamlit. However, whenever I would click on a marker, the "last_object_clicked" part of the dictionary would not update. It would only update after you then clicked on the map (but not a marker) or move the map and "last_clicked" or "bounds" respectively would update along with "last_object_clicked", re-running the streamlit app. This error can be reproduced on your /examples/interactive_app.py. 

I found that in the console when you only click the marker it throws an "Uncaught Type Error: t.drawnItems.toGeoJSON is not a function" relating to the onLayerClick function here. It seems as though there is some error with drawnItems, and so the "debouncedUpdateComponentValue()" is never run, which is why clicking the marker doesn't update until the debounced function is run elsewhere.

I've separated the two functions out and made a specific function for clicking the markers, onObjectClick. I'm not sure if this fixes the error thrown when using the drawing tool because I am not using that tool for my project, but this does make it so that when clicking a marker the last_object_clicked updates and works fine.